requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
accelerate=0.22.0=pypi_0
appdirs=1.4.4=pypi_0
asttokens=2.0.5=pyhd3eb1b0_0
attrs=23.1.0=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.01.10=h06a4308_0
certifi=2022.12.7=py38h06a4308_0
cffi=1.15.1=py38h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
comm=0.1.2=py38h06a4308_0
cryptography=41.0.2=py38h22a60cf_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
datasets=2.14.5=pypi_0
debugpy=1.5.1=py38h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
docker-pycreds=0.4.0=pypi_0
entrypoints=0.4=py38h06a4308_0
executing=0.8.3=pyhd3eb1b0_0
ffmpeg=4.3=hf484d3e_0
filelock=3.9.0=py38h06a4308_0
freetype=2.12.1=h4a9f257_0
fsspec=2023.9.0=pypi_0
giflib=5.2.1=h5eee18b_3
gitdb=4.0.10=pypi_0
gitpython=3.1.35=pypi_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py38heeb90bb_0
gnutls=3.6.15=he1e5248_0
idna=3.4=py38h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46305
ipykernel=6.19.2=py38hb070fc8_0
ipython=8.8.0=py38h06a4308_0
jedi=0.18.1=py38h06a4308_1
jinja2=3.1.2=py38h06a4308_0
jpeg=9e=h5eee18b_1
jupyter_client=7.4.8=py38h06a4308_0
jupyter_core=5.1.1=py38h06a4308_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.7.2.10=0
libcurand=10.3.3.141=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.1=py38h7f8727e_0
matplotlib-inline=0.1.6=py38h06a4308_0
mkl=2023.1.0=h213fc3f_46343
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.6=py38h417a72b_1
mkl_random=1.2.2=py38h417a72b_1
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py38h06a4308_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=py38h06a4308_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py38h06a4308_0
numpy=1.24.3=py38hf6e8229_1
numpy-base=1.24.3=py38h060ed82_1
openh264=2.1.1=h4ff587b_0
openssl=3.0.10=h7f8727e_2
packaging=22.0=py38h06a4308_0
parso=0.8.3=pyhd3eb1b0_0
pathtools=0.1.2=pypi_0
peft=0.5.0=pypi_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.4.0=py38h6a678d5_0
pip=23.2.1=py38h06a4308_0
platformdirs=2.5.2=py38h06a4308_0
prompt-toolkit=3.0.36=py38h06a4308_0
protobuf=4.24.3=pypi_0
psutil=5.9.5=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=23.2.0=py38h06a4308_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.17=h955ad1f_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=2.0.1=py3.8_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-mutex=1.0=cuda
pyyaml=6.0.1=pypi_0
pyzmq=23.2.0=py38h6a678d5_0
readline=8.2=h5eee18b_0
requests=2.31.0=py38h06a4308_0
scikit-learn=1.3.0=pypi_0
sentry-sdk=1.30.0=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=68.0.0=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.2.0=pyhd3eb1b0_0
sympy=1.11.1=py38h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tk=8.6.12=h1ccaba5_0
torchaudio=2.0.2=py38_cu118
torchtriton=2.0.0=py38
torchvision=0.15.2=py38_cu118
tornado=6.2=py38h5eee18b_0
traitlets=5.7.1=py38h06a4308_0
transformers=4.33.1=pypi_0
typing_extensions=4.7.1=py38h06a4308_0
urllib3=1.26.16=py38h06a4308_0
wandb=0.15.10=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
wheel=0.38.4=py38h06a4308_0
xz=5.4.2=h5eee18b_0
zeromq=4.3.4=h2531618_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0